+++ build.gradle
@@ -6,100 +6,22 @@ plugins {
 
 group 'io.github.amrjlg'
 version '1.3-SNAPSHOT'
-archivesBaseName = "commons"
+archivesBaseName = "common"
 
 
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation 'org.apache.commons:commons-lang3:3.6'
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.0'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
-}
-
-test {
-    useJUnitPlatform()
-}
-
-tasks.withType(JavaCompile) {
-    options.encoding = "UTF-8"
-}
-// for more information https://docs.gradle.org/current/javadoc/org/gradle/external/javadoc/StandardJavadocDocletOptions.html
-tasks.withType(Javadoc) {
-    options {
-        encoding "UTF-8"
-        charSet "UTF-8"
-        author true
-        version true
-        // 需要https才能生效
-        links "https://docs.oracle.com/javase/8/docs/api"
-        title "common"
-    }
-
-}
-
-
-task javadocJar(type: Jar, dependsOn: javadoc) {
-    classifier = 'javadoc'
-    from javadoc
-}
-
-task sourcesJar(type: Jar) {
-    classifier = 'sources'
-    from sourceSets.main.allSource
-}
-
-artifacts {
-    archives javadocJar, sourcesJar
-}
-
-signing {
-    sign configurations.archives
-}
-
-uploadArchives {
-    repositories {
-        mavenDeployer {
-            beforeDeployment { MavenDeployment deployment -> signing.signPom(deployment) }
-
-            repository(url: "https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/") {
-                authentication(userName: "${ossrhUsername}", password: "${ossrhPassword}")
+configure(subprojects) {
+    project ->
+        if (project.name != 'buildSrc'){
+            task javadocJar(type: Jar, dependsOn: javadoc) {
+                classifier = 'javadoc'
+                from javadoc
             }
 
-            snapshotRepository(url: "https://s01.oss.sonatype.org/content/repositories/snapshots/") {
-                authentication(userName: "${ossrhUsername}", password: "${ossrhPassword}")
+            task sourcesJar(type: Jar) {
+                classifier = 'sources'
+                from sourceSets.main.allSource
             }
-
-            pom.project {
-                name 'common utils.'
-                packaging 'jar'
-                // optionally artifactId can be defined here
-                description 'common utils.'
-                url 'https://github.com/amrjlg/common'
-
-                scm {
-                    connection 'https://github.com/amrjlg/common.git'
-                    developerConnection 'https://github.com/amrjlg/common.git'
-                    url 'https://github.com/amrjlg/common'
-                }
-
-                licenses {
-                    license {
-                        name 'The Apache License, Version 2.0'
-                        url 'https://www.apache.org/licenses/LICENSE-2.0.txt'
-                    }
-                }
-
-                developers {
-                    developer {
-                        id 'amrjlg'
-                        name 'amrjlg'
-                        email '2899297045@qq.com'
-                    }
-                }
-            }
+            apply from:'../buildSrc/sigin.gradle'
         }
-    }
+
 }
