+++ buildSrc/sigin.gradle
@@ -1,0 +1,91 @@
+/*
+ * Copyright (c) 2021-2021 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+
+apply plugin:'maven'
+apply plugin:'signing'
+
+
+tasks.withType(JavaCompile) {
+    options.encoding = "UTF-8"
+}
+// for more information https://docs.gradle.org/current/javadoc/org/gradle/external/javadoc/StandardJavadocDocletOptions.html
+tasks.withType(Javadoc) {
+    options {
+        encoding "UTF-8"
+        charSet "UTF-8"
+        author true
+        version true
+        // 需要https才能生效
+        links "https://docs.oracle.com/javase/8/docs/api"
+        title "common"
+    }
+
+}
+
+
+artifacts {
+    archives javadocJar, sourcesJar
+}
+
+signing {
+    sign configurations.archives
+}
+
+uploadArchives {
+    repositories {
+        mavenDeployer {
+            beforeDeployment { MavenDeployment deployment -> signing.signPom(deployment) }
+
+            repository(url: "https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/") {
+                authentication(userName: "${ossrhUsername}", password: "${ossrhPassword}")
+            }
+
+            snapshotRepository(url: "https://s01.oss.sonatype.org/content/repositories/snapshots/") {
+                authentication(userName: "${ossrhUsername}", password: "${ossrhPassword}")
+            }
+
+            pom.project {
+                name project.name
+                packaging 'jar'
+                // optionally artifactId can be defined here
+                description project.description
+                url 'https://github.com/amrjlg/common'
+
+                scm {
+                    connection 'https://github.com/amrjlg/common.git'
+                    developerConnection 'https://github.com/amrjlg/common.git'
+                    url 'https://github.com/amrjlg/common'
+                }
+
+                licenses {
+                    license {
+                        name 'The Apache License, Version 2.0'
+                        url 'https://www.apache.org/licenses/LICENSE-2.0.txt'
+                    }
+                }
+
+                developers {
+                    developer {
+                        id 'amrjlg'
+                        name 'amrjlg'
+                        email '2899297045@qq.com'
+                    }
+                }
+            }
+        }
+    }
+}
