+++ exception/src/main/java/io/github/amrjlg/exception/DefaultExceptionHandler.java
@@ -1,0 +1,32 @@
+/*
+ *  Copyright (c) 2021-2021 the original author or authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package io.github.amrjlg.exception;
+
+/**
+ * @author jiang
+ */
+public class DefaultExceptionHandler extends AbstractExceptionHandler<Exception> {
+
+    public DefaultExceptionHandler() {
+        super(Exception.class::isAssignableFrom);
+    }
+
+    @Override
+    public String resolve(Exception e) {
+        return e.getMessage();
+    }
+}
