+++ exception/src/main/java/io/github/amrjlg/exception/ExceptionHandler.java
@@ -1,0 +1,39 @@
+/*
+ *  Copyright (c) 2021-2021 the original author or authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package io.github.amrjlg.exception;
+
+import io.github.amrjlg.api.Handler;
+import io.github.amrjlg.api.Support;
+
+/**
+ * exception handler
+ * <p>
+ * 2021/3/4
+ *
+ * @author jiang
+ **/
+public interface ExceptionHandler<T extends Exception> extends Handler<String, T>, Support<Exception> {
+    /**
+     * 处理异常
+     *
+     * @param e 异常
+     * @return 异常处理结果
+     */
+    @Override
+    String resolve(T e);
+
+}
