+++ id/src/main/java/io/github/amrjlg/id/Snowflake.java
