--- src/test/java/io/github/amrjlg/steam/ArrayStreamTest.java
@@ -1,14 +1,0 @@
-package io.github.amrjlg.steam;
-
-import org.junit.jupiter.api.Test;
-
-import java.util.Arrays;
-
-public class ArrayStreamTest {
-
-    @Test
-    public void stream(){
-        int sum = Arrays.stream(new double[]{1, 2, 3, 4}).mapToInt(d -> (int) d).sum();
-
-    }
-}
