+++ util/src/main/java/io/github/amrjlg/utils/ArrayUtil.java
