+++ util/src/main/java/io/github/amrjlg/utils/ByteUtil.java
