+++ util/src/main/java/io/github/amrjlg/utils/HexUtil.java
