+++ util/src/main/java/io/github/amrjlg/utils/NetUtil.java
