+++ util/src/main/java/io/github/amrjlg/utils/StringUtil.java
